Registration
I've added the registration functionality. The user can register in one step using
a simple regisration (without email confirmation). To limit the risk of a random person
destroying me all the content, the newly register person can only add entries and edit
etries added by he/she. They are not allowed to delete entries or edit other people entries.

I've also added the paper button on submit. It is possible to click it, but it's not possible
now to submit the form using ENTER.